TriggerCampus/requirements.txt
matplotlib
pandas
scikit-learn
azure-storage-blob
requests
streamlit
numpy==1.19.3
pyodbc
azure-functions